LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStreamTrack-transfer-video.https-expected.txt

CONSOLE MESSAGE: DataCloneError: The object can not be cloned.


Harness Error (FAIL), message = DataCloneError: The object can not be cloned.

TIMEOUT MediaStreamTrack transfer to iframe Test timed out

Harness Error (FAIL), message = DataCloneError: The object can not be cloned.

TIMEOUT MediaStreamTrack transfer to iframe Test timed out
